Case 1 Mapped ERD
A big company has decided to store information about its projects and 
employees in a database. The company has wisely chosen to hire you as a 
database designer. Prepare an E-R diagram for this Company according to 
The following Description: 
• The company has a number of employees each employee has SSN, Birth 
Date, Gender and Name which represented as Fname and Lname. 
• The company has a set of departments each department has a set of attributes 
DName, DNUM (unique) and locations. 
• Employees work in several projects each project has Pname, PNumber as an 
identifier, Location and City. 
• Each employee may have a set of dependent; each dependent has Dependent 
Name (unique), Gender, and Birth Date. 
Note: if the employee left the company no needs to store his dependents info 
• For each Department, there is always one employee assigned to manage that 
Department and each manager has a hiring Date  
• Department may have employees but employee must work on Only One 
department 
• Each department may have a set of projects and each project must assigned 
to one department 
• Employees work in several projects and each project has several employees 
and each employee has a number of working hours in each project 
• Each employee has a supervisor